requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
tensorflow>=2.10.0
pillow>=8.0.0
numpy>=1.20.0
opencv-python-headless>=4.5.0
scikit-learn>=1.0.0
requests>=2.25.0
werkzeug>=2.0.0
gunicorn>=21.0.0 